refactor: 연도와 학기를 Integer로 다루도록 변경

+++ server/src/main/java/wap/web2/server/project/repository/ProjectRepository.java
@@ -11,7 +11,7 @@ import wap.web2.server.project.entity.Project;
 public interface ProjectRepository extends JpaRepository<Project, Long> {
 
     @Query(value = "select p from Project p where p.projectYear = :year and p.semester = :semester")
-    List<Project> findProjectsByYearAndSemester(@Param("year") Long year, @Param("semester") Long semester);
+    List<Project> findProjectsByYearAndSemester(@Param("year") Integer year, @Param("semester") Integer semester);
 
     @Query("SELECT b FROM Project b WHERE b.projectId = :projectId AND b.user.id = :userId")
     Optional<Project> findByProjectIdAndUser(@Param("projectId") Long projectId, @Param("userId") Long userId);
@@ -21,7 +21,7 @@ public interface ProjectRepository extends JpaRepository<Project, Long> {
     int voteByProjectId(@Param("projectId") Long projectId);
 
     @Query("SELECT p FROM Project p WHERE p.projectYear = :year AND p.semester = :semester ORDER BY p.projectId DESC")
-    List<Project> findProjectsByYearAndSemesterOrderByProjectIdDesc(@Param("year") Long year,
-                                                                    @Param("semester") Long semester);
+    List<Project> findProjectsByYearAndSemesterOrderByProjectIdDesc(@Param("year") Integer year,
+                                                                    @Param("semester") Integer semester);
     
 }
